update comments and articles

+++ app/Http/Controllers/CommentController.php
@@ -2,15 +2,24 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use App\Models\Article;
 
 class CommentController extends Controller
 {
+    public function __construct() {
+        $this->middleware('auth');
+    }
     public function delete($id) {
         $comment = Comment::find($id);
-        $comment->delete();
-        return back()->with("cm-delete", "Comment deleted");
+        if(Gate::allows("comment-delete", $comment)) {
+            $comment->delete();
+            return back()->with('cm-delete', "Comment deleted");
+        } else {
+            return back()->with('unanthorized', "Unauthorized");
+        }
     }
     public function add() {
         $validator = validator(request()->all(), [
@@ -23,7 +32,16 @@ class CommentController extends Controller
         $comment = new Comment;
         $comment->content = request()->content;
         $comment->article_id = request()->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
         return back()->with("cm-created", "Comment created");
+    }
+    public function view($id) {
+        $articles = Article::find($id);
+        $comment = Comment::find($id);
+        return view("comments.view", [
+            'comments' => $comment,
+            'article' => $articles,
+        ]);
     }
 }
